requirements.txt
streamlit==1.37.0
pandas==2.2.2
plotly==5.21.0
google-genai==0.5.4
python-dotenv==1.0.1